+++ test/test.c
@@ -379,6 +379,10 @@ test_find_function(void) {
         && lwjson_get_val_string_length(token) == 11
         && strncmp(token->u.str.token_value, "\\t\\u1234abc", 11) == 0);
 
+    /* Check string compare */
+    RUN_TEST((token = lwjson_find_ex(&lwjson, NULL, "my_obj.arr.#.#.my_key")) != NULL
+                && lwjson_string_compare(token, "my_text"));
+
 #undef RUN_TEST
 
     /* Call this once JSON usage is finished */